deps/v8/test/inspector/debugger/step-into-next-script-expected.txt

Debugger breaks in next script after stepOut from previous one.

Running test: testStepOut
test (foo.js:12:2)
(anonymous) (:0:0)

(anonymous) (:0:6)

(anonymous) (timeout1.js:0:8)

foo (timeout2.js:19:2)

(anonymous) (timeout3.js:0:8)


Running test: testStepOver
(anonymous) (:0:0)

test (foo.js:12:2)
(anonymous) (:0:0)

test (foo.js:13:0)
(anonymous) (:0:0)

(anonymous) (:0:6)

(anonymous) (timeout1.js:0:8)

(anonymous) (timeout1.js:0:35)

foo (timeout2.js:19:2)

foo (timeout2.js:19:12)

(anonymous) (timeout3.js:0:8)

(anonymous) (timeout3.js:0:35)


Running test: testStepInto
(anonymous) (:0:0)

test (foo.js:9:2)
(anonymous) (:0:0)

test (foo.js:10:2)
(anonymous) (:0:0)

test (foo.js:11:2)
(anonymous) (:0:0)

test (foo.js:12:2)
(anonymous) (:0:0)

test (foo.js:13:0)
(anonymous) (:0:0)

(anonymous) (:0:6)

(anonymous) (timeout1.js:0:8)

(anonymous) (timeout1.js:0:35)

foo (timeout2.js:19:2)

foo (timeout2.js:19:12)

(anonymous) (timeout3.js:0:8)

(anonymous) (timeout3.js:0:35)
